% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRatools.R
\docType{package}
\name{mRatools-package}
\alias{mRatools}
\alias{mRatools-package}
\title{mRatools: Microbiological risk assessment tools}
\description{
This package is designed to make it easy to install and load multiple microbiological risk assessment based R packages in a single step.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Mayooran1987/mRatools}
  \item Report bugs at \url{https://github.com/Mayooran1987/mRatools/issues}
}

}
\author{
\strong{Maintainer}: Mayooran Thevaraja \email{mayooran@eng.jfn.ac.lk}

Authors:
\itemize{
  \item Kondaswamy Govindaraju \email{k.govindaraju@massey.ac.nz}
  \item Mark Bebbington \email{m.bebbington@massey.ac.nz}
}

}
\keyword{internal}
