% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{mRatools_conflicts}
\alias{mRatools_conflicts}
\title{Conflicts between the mRatools and other packages}
\usage{
mRatools_conflicts(only = NULL)
}
\arguments{
\item{only}{Set this to a character vector to restrict to conflicts only
with these packages.}
}
\description{
This function lists all the conflicts between packages in the mRatools
and other packages that you have loaded.
}
\details{
There are no conflicts between the mRatools and other packages at this stage.
}
\examples{
mRatools_conflicts()
}
