% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mRatools_packages}
\alias{mRatools_packages}
\title{List all packages in the mRatools}
\usage{
mRatools_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include mRatools in the list?}
}
\description{
List all packages in the mRatools
}
\examples{
mRatools_packages()
}
